Relocate tests to dedicated folder. Ignore tests, stories, and ide configs in npmignore. Run tests as part of amplify deploy
