+++ __tests__/helpers/BasicPointsAndBadges/ConnectExternalAccountActivity.test.ts
@@ -1,4 +1,4 @@
-import { awardConnectExternalAccountActivityPoints, ConnectExternalAccountActivity } from "./ConnectExternalAccountActivity";
+import { awardConnectExternalAccountActivityPoints, ConnectExternalAccountActivity } from "../../../src/helpers/BasicPointsAndBadges/ConnectExternalAccountActivity";
 
 jest.mock('@careevolution/mydatahelps-js', () => {
