+++ amplify.yml
@@ -7,6 +7,7 @@ frontend:
         - npm ci
     build:
       commands:
+        - npm run test
         - npm run build-storybook
   artifacts:
     baseDirectory: /storybook-static/
