+++ package.json
@@ -59,8 +59,7 @@
     "@storybook/addon-webpack5-compiler-babel": "^3.0.3",
     "@storybook/react": "^8.3.5",
     "@storybook/react-webpack5": "^8.3.5",
-    "@storybook/testing-library": "^0.2.2",
-    "@types/jest": "^29.5.13",
+    "@types/jest": "^29.5.14",
     "@types/lodash": "^4.17.7",
     "@types/markdown-it": "^12.2.3",
     "@types/react": "^18.0.33",
